Adds BroadcastStats
